# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Shine - Harvard International Students App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    put:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/onboard:
    post:
      operationId: onboardUser
      tags: [users]
      summary: Complete onboarding for a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardUserBody"
      responses:
        "201":
          description: User created/onboarded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /teams:
    get:
      operationId: listTeams
      tags: [teams]
      summary: List all teams
      responses:
        "200":
          description: Teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
    post:
      operationId: createTeam
      tags: [teams]
      summary: Create a new team
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamBody"
      responses:
        "201":
          description: Team created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
  /teams/me:
    get:
      operationId: getMyTeam
      tags: [teams]
      summary: Get current user's team
      responses:
        "200":
          description: My team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "404":
          description: Not on a team
  /teams/{id}:
    get:
      operationId: getTeam
      tags: [teams]
      summary: Get a team by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "404":
          description: Not found
  /teams/{id}/join:
    post:
      operationId: joinTeam
      tags: [teams]
      summary: Join a team
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Joined team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"

  # ─── Submissions ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: string
        displayName:
          type: string
        pseudonym:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        timezone:
          type: ["string", "null"]
        concentration:
          type: ["string", "null"]
        role:
          type: string
          enum: [student, mentor, admin]
        phase:
          type: string
          enum: [pre_arrival, on_campus]
        comfortSpeaking:
          type: integer
        comfortAsking:
          type: integer
        comfortMeeting:
          type: integer
        onboardingCompleted:
          type: boolean
        totalPoints:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sessionId
        - displayName
        - role
        - phase
        - comfortSpeaking
        - comfortAsking
        - comfortMeeting
        - onboardingCompleted
        - totalPoints
        - createdAt
    UpdateUserBody:
      type: object
      properties:
        displayName:
          type: string
        pseudonym:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        timezone:
          type: ["string", "null"]
        concentration:
          type: ["string", "null"]
        phase:
          type: string
          enum: [pre_arrival, on_campus]
        comfortSpeaking:
          type: integer
        comfortAsking:
          type: integer
        comfortMeeting:
          type: integer
    OnboardUserBody:
      type: object
      properties:
        sessionId:
          type: string
        displayName:
          type: string
        pseudonym:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        timezone:
          type: ["string", "null"]
        concentration:
          type: ["string", "null"]
        phase:
          type: string
          enum: [pre_arrival, on_campus]
        comfortSpeaking:
          type: integer
        comfortAsking:
          type: integer
        comfortMeeting:
          type: integer
      required:
        - sessionId
        - displayName
...
    Team:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        totalPoints:
          type: integer
        memberCount:
          type: integer
        members:
          type: array
          items:
            $ref: "#/components/schemas/TeamMember"
        completedMissions:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - totalPoints
        - memberCount
        - members
        - completedMissions
        - createdAt
    TeamMember:
      type: object
      properties:
        userId:
          type: integer
        displayName:
          type: string
        country:
          type: ["string", "null"]
        totalPoints:
          type: integer
      required:
        - userId
        - displayName
        - totalPoints
    CreateTeamBody:
      type: object
      properties:
        name:
          type: string
      required:
        - name
...