# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shine - Harvard International Students App API
 * OpenAPI spec version: 0.1.0
 */
import type { UserPhase } from "./userPhase";
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  sessionId: string;
  displayName: string;
  /** @nullable */
  pseudonym?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  timezone?: string | null;
  /** @nullable */
  concentration?: string | null;
  role: UserRole;
  phase: UserPhase;
  comfortSpeaking: number;
  comfortAsking: number;
  comfortMeeting: number;
  onboardingCompleted: boolean;
  totalPoints: number;
  createdAt: Date;
}